requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==4.2.19
djangorestframework==3.15.2
djongo==1.2.31
dnspython==2.7.0
mongoengine==0.29.1
pymongo==4.11.1
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
